code : BOJ 5618 B2 공약수
